third_party/WebKit/LayoutTests/fast/dom/MutationObserver/clear-transient-without-delivery-expected.txt

Transient registrations should be cleared even without delivery.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutationsDelivered is false
PASS successfullyParsed is true

TEST COMPLETE
